+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1.4 ##
+
+*   Add `AddressableLabel`. This only works if you have `Addressable` installed in your project.
+
 ## 1.1.3 ##
 
 *   `AnimatorParam` no longer offers `null` value